🚚 Rename: 파일이름 index.tsx로 변경 #22
