+++ src/components/auth-Intro-text/index.tsx
